Change event published notification logic
Changed event publish notification to be sent to every child of the
project. Previously there was only one notification sent per guardian.
