+++ events/admin.py
@@ -2,11 +2,7 @@ from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
 from parler.admin import TranslatableAdmin
 
-from users.models import Guardian
-
 from .models import Event, Occurrence
-from .notifications import NotificationType
-from .utils import send_event_notifications_to_guardians
 
 
 class OccurrencesInline(admin.StackedInline):
@@ -34,12 +30,6 @@ class EventAdmin(TranslatableAdmin):
     def publish(self, request, queryset):
         for obj in queryset:
             obj.publish()
-            guardians = Guardian.objects.filter(
-                children__project=obj.project
-            ).distinct()
-            send_event_notifications_to_guardians(
-                obj, NotificationType.EVENT_PUBLISHED, guardians
-            )
         self.message_user(request, _("%s successfully published.") % queryset.count())
 
     publish.short_description = _("Publish selected events")
