+++ events/consts.py
@@ -1,0 +1,4 @@
+class NotificationType:
+    EVENT_PUBLISHED = "event_published"
+    OCCURRENCE_ENROLMENT = "occurrence_enrolment"
+    OCCURRENCE_UNENROLMENT = "occurrence_unenrolment"
