+++ events/signals.py
@@ -10,24 +10,20 @@ from events.utils import send_event_notifications_to_guardians
 @receiver(post_save, sender=Enrolment, dispatch_uid="send_enrolment_email")
 def send_enrolment_email(instance, created, **kwargs):
     if created:
-        guardians = instance.child.guardians.all()
         send_event_notifications_to_guardians(
             instance.occurrence.event,
             NotificationType.OCCURRENCE_ENROLMENT,
-            guardians,
-            child=instance.child,
+            instance.child,
             occurrence=instance.occurrence,
         )
 
 
 @receiver(post_delete, sender=Enrolment, dispatch_uid="send_unenrolment_email")
 def send_unenrolment_email(instance, **kwargs):
-    guardians = instance.child.guardians.all()
     send_event_notifications_to_guardians(
         instance.occurrence.event,
         NotificationType.OCCURRENCE_UNENROLMENT,
-        guardians,
-        child=instance.child,
+        instance.child,
         occurrence=instance.occurrence,
     )
